// src/app/user-stats/models/stats.ts
export class Stats {
    username: string;
    sport: string;
    Benefit: number;
    StakeAverage: number;
    Yield: number;
    OddAverage: number;
    TotalPicks: number;
    OddAccumulator: number;
    TotalStaked: number;
    Wins: number;
    Voids: number;
    Lost: number;
}